+++ image_edit/algorithms.h
@@ -147,12 +147,12 @@ void histogram_equalization_bw(
 
 /* Image Convolution with one mask */
 void image_convolution_bw(
-    unsigned char ***image_in, unsigned char ***image_out, float **w, int size, double factor, double bias
+    unsigned char ***image_in, unsigned char ***image_out, double **w, int size, double factor, double bias
     );
 
 /* Image Convolution with 2 mask, X-Axis and Y-Axis */
 void image_convolution_2d_bw(
-    unsigned char ***image_in, unsigned char ***image_out, float **wx, float **wy, int size, double factor, double bias
+    unsigned char ***image_in, unsigned char ***image_out, double **wx, double **wy, int size, double factor, double bias
     );
 
 /* Image rotation (clockwise) */
@@ -307,14 +307,14 @@ void histogram_equalization_yuv_color(
 void image_convolution_color(
     unsigned char ***image_in_r, unsigned char ***image_in_g, unsigned char ***image_in_b,
     unsigned char ***image_out_r, unsigned char ***image_out_g, unsigned char ***image_out_b,
-    float **w, int size, double factor, double bias
+    double **w, int size, double factor, double bias
     );
 
 /* Image Convolution with 2 mask, X-Axis and Y-Axis */
 void image_convolution_2d_color(
     unsigned char ***image_in_r, unsigned char ***image_in_g, unsigned char ***image_in_b,
     unsigned char ***image_out_r, unsigned char ***image_out_g, unsigned char ***image_out_b,
-    float **wx, float **wy, int size, double factor, double bias
+    double **wx, double **wy, int size, double factor, double bias
     );
 
 /* Image rotation (clockwise) */
